+++ src/main/java/camp/woowak/lab/web/dto/request/store/StoreRegistrationRequest.java
@@ -1,4 +1,4 @@
-package camp.woowak.lab.store.service.dto;
+package camp.woowak.lab.web.dto.request.store;
 
 import java.time.LocalDateTime;
